Removed SQL link from sidebar and updated workspace settings
The SQL link was commented out from the ‘handleSidebar.js’ file, removing it from the sidebar navigation. The ‘contact.html' link was also adjusted in the same file. This commit also includes updates to the 'workspace.xml' under the .idea directory, such as rearranging the order of templates and updating the local task counter. Some old commit messages were deleted for simplicity.